astro/requirements.txt
asgiref==3.8.1
Django==5.1.4
pyswisseph==2.10.3.2
pytz==2024.2
sqlparse==0.5.3
typing_extensions==4.12.2
